feat: cleared main and moved old main to another file
